rn: Move the account info style into its component
